move module_requests
