+++ playground/src/pages/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import BaseInfoHeader from "@/components/base_info_header";
-import ModuleRequests from "@/components/module_requests";
+import ModuleRequests from "@/components/view_mode/module_requests";
 import { useBaseInfo } from "@/hooks/base_info_context";
